webtop: edit init script

+++ webtop/custom-cont-init.d/init.sh
@@ -1,24 +1,38 @@
 #!/bin/bash
 
-pushd /var/lib/Pincer/bin
-chmod +x ./setupraw.sh
-su abc -c './setupraw.sh'
-popd
+if [[ ! -f /home/anubis/.davinci.conf ]]
+then
 
-mv /var/lib/content.zip /home/anubis/content.zip
-pushd /home/anubis
-mkdir study_content
-mv content.zip ./study_content/
-popd
+  # install extension from .vsix file
+  pushd /var/lib/Pincer/bin
+  chmod +x ./setupraw.sh
+  su abc -c './setupraw.sh'
+  popd
+fi
 
-pushd /home/anubis/study_content
-unzip content.zip
-rm -rf content.zip
-popd
+if [[ ! -d /home/anubis/study_content ]]
+then
 
-chown -c -R abc:abc /home/anubis/study_content
+  # copy over contents of the study
+  mv /var/lib/content.zip /home/anubis/content.zip
+  pushd /home/anubis
+  mkdir study_content
+  mv content.zip ./study_content/
+  popd
 
-mkdir /home/anubis/.config/autostart
+  pushd /home/anubis/study_content
+  unzip content.zip
+  rm -rf content.zip
+  popd
 
-mv /home/anubis/custom-cont-init.d/code.desktop /home/anubis/.config/autostart/
+  chown -c -R abc:abc /home/anubis/study_content
+fi
+
+if [[ ! -d /home/anubis/.config/autostart ]]
+then
+
+  # create start up scripts for xfce
+  mkdir /home/anubis/.config/autostart
+  mv /home/anubis/custom-cont-init.d/code.desktop /home/anubis/.config/autostart/
+fi
